Clean up functions, remove some repetitive code
